requirements.txt
flask
flask-smorest
sqlalchemy
flask-sqlalchemy
python-dotenv